071の途中

+++ ABC/169/C.cpp
@@ -3,7 +3,7 @@ using namespace std;
 
 int main() {
   int64_t a;
-  double b;
+  long double b;
   cin >> a >> b;
 
   int64_t ans = static_cast<int64_t>(a * b);
