+++ AGC/012/A.cpp
@@ -2,15 +2,15 @@
 using namespace std;
 
 int main() {
-  int n;
+  int64_t n;
   cin >> n;
-  vector<int> a(3 * n);
+  vector<int64_t> a(3 * n);
   for (int i = 0; i < n; ++i) {
     cin >> a.at(3 * i) >> a.at(3 * i + 1) >> a.at(3 * i + 2);
   }
   sort(a.begin(), a.end());
 
-  int sum = 0;
+  int64_t sum = 0;
   for (int i = 0; i < n; ++i) {
     sum += a.at(n + 2 * i);
   }
